fix: missing package and wrong svelte code
The package @tiptap/extension-bubble-menu was missing in the instructions but used in the example.

Also, the code was missing the opening script tag and the onMount import. I also moved the onDestroy to the return function of the onMount since it does the same.

I would also recommend for the future to use a `<script lang="ts">` tag and write it in TypeScript.